+++ crates/matrix-sdk-crypto/CHANGELOG.md
@@ -35,6 +35,8 @@ All notable changes to this project will be documented in this file.
   ([#6044](https://github.com/matrix-org/matrix-rust-sdk/pull/6044))
 - Add support for MSC4385.
   - Add new method `OlmMachine::push_secret_to_verified_devices`.
+  - Pushed secrets that we receive from verified devices are added to the
+    secrets inbox.
 
 ### Refactor